renamefiles: Use print function, improve error handling.

+++ renamefiles.py
@@ -1,7 +1,13 @@
+from __future__ import print_function
 import os
 import sys
 
+# TODO improve arg parsing (use a library)
 renames = sys.argv[1].strip()
+force = sys.argv[2].strip() == '-f'
+
+if force:
+    print("-f detected: will overwrite files")
 
 with open(renames, "r") as f:
     while True:
@@ -13,8 +19,18 @@ with open(renames, "r") as f:
         a = args[0].strip()
         b = args[1].strip()
 
-        print a,
-        print " --> ",
-        print b
+        print(a, end = " ")
+        print(" --> ", end = " ")
+        print(b)
 
-        os.rename(a, b)
+        try:
+            os.rename(a, b)
+        except FileNotFoundError as e:
+            print(e)
+        except FileExistsError as e:
+            if force:
+                print("Overwriting {}".format(b))
+                os.remove(b)
+                os.rename(a, b)
+            else:
+                print(e)
